+++ packages/valory/skills/termination_abci/models.py
@@ -29,7 +29,7 @@ from packages.valory.skills.abstract_round_abci.models import Requests as BaseRe
 from packages.valory.skills.abstract_round_abci.models import (
     SharedState as BaseSharedState,
 )
-from packages.valory.skills.termination_abci.behaviours import TerminationAbciApp
+from packages.valory.skills.termination_abci.rounds import TerminationAbciApp
 
 
 class SharedState(BaseSharedState):
